api_ingestion/requirements.txt
fastapi
uvicorn[standard]
langchain
openai
pydantic
psycopg2-binary   # for PostgreSQL + pgvector
pgvector
sqlalchemy
python-dotenv
